changed newfile.py not to write over scripts

+++ newfile.py
@@ -9,10 +9,14 @@ cookie = {
     'session':'53616c7465645f5f3629edbaeb85a2b8c366f7c09adb785325acf4473fb53689f90437954735668c6ae0aa7af0ecf6d38da5c9db1222a6b1e66f6fd6d0b75825'}
 for folder in ('inputs','scripts'):
     if not os.path.isdir(f'/{folder}/{year}'):
-        os.makedirs(f'{folder}/{year}')
+        try:
+            os.makedirs(f'{folder}/{year}')
+        except FileExistsError:
+            pass
 r = requests.get(f'https://adventofcode.com/{year}/day/{day}/input',cookies=cookie)
 with open(f'inputs/{year}/{day}.txt','w') as f:
     f.write(r.content.decode())
-lines = [str(line).replace('year',f'{year}').replace('day',f'{day}') for line in open('template.txt').readlines()]
-with open(f'scripts/{year}/{day}.py','w') as f:
-    f.write(''.join(lines))
+if not os.path.isfile(f'scripts/{year}/{day}.py'):
+    lines = [str(line).replace('year',f'{year}').replace('day',f'{day}') for line in open('template.txt').readlines()]
+    with open(f'scripts/{year}/{day}.py','w') as f:
+        f.write(''.join(lines))
